Commonized CheckSum Builder.
