+++ BuildScripts/Deb.py
@@ -10,6 +10,8 @@ import os
 import re
 import subprocess
 
+import Common
+
 Import('envBase')
 
 debEnv = envBase.Clone()
@@ -123,16 +125,7 @@ debTarget = debEnv.Command(
 # Checksum
 ###
 
-def Checksum(target, source, env):
-    with io.open(str(source[0]), 'rb') as inFile:
-        readFile = inFile.read()
-        h = hashlib.sha256(readFile)
-        hash = h.hexdigest()
-
-    with io.open(str(target[0]), 'w', encoding='utf-8') as outFile:
-        outFile.write(unicode(hash))
-
-debEnv.Append(BUILDERS = {"Checksum" : Builder(action=Checksum)})
+debEnv.Append(BUILDERS = {"Checksum" : Builder(action=Common.Checksum)})
 
 checksumTarget = debEnv.Checksum(
     target = os.path.join(debEnv['DEBIAN_INSTALL_DIR'], 'bin', 'chaskis.deb.sha256'),
